Walkcycle, block shadows, a few fruit, and pillar fix
metasprites.nss:
Idle, Blink, and Walkcycle for left and right built. (some old metasprites deleted.)
AppleLow, AppleHigh; PearLow and PearHigh fruit metasprites built.

BlockShadow1 through 5. The number denotes which to show based on distance block is from it. If falling block is 1 step away, BlockShadow1. If 5 or more steps away, BlockShadow5.
If BlockShadow is overlapped by a portion of the block, should not be rendered (for scanline conservation)

StarlitStablesBG.chr: pillars are a uniform size now.
StarlitStablesSPR.pal: corrected.